:recycle: Refactor code
