+++ Website/Sources/Pages/CodeOfConduct.swift
@@ -1,6 +1,6 @@
 import Ignite
 
-enum CodeOfConductSectionType: String, CaseIterable {
+enum CodeOfConductSectionType: String, StringEnum {
   case communityParticipation = "community_participation"
   case harassmentFree = "harassment_free"
   case forSpeakers = "for_speakers"
@@ -25,30 +25,8 @@ struct CodeOfConduct: StaticLayout {
   var body: some HTML {
     MainNavigationBar(path: generatePath(language:), language: language)
 
-    Text(String(forKey: "code_of_conduct", language: language))
-      .horizontalAlignment(.center)
-      .font(.title1)
-      .fontWeight(.bold)
-      .foregroundStyle(.bootstrapPurple)
-      .padding(.top, 140)
-
-    ForEach(CodeOfConductSectionType.allCases) { sectionType in
-      Section {
-        Text(String(forKey: sectionType.rawValue, language: language))
-          .horizontalAlignment(.center)
-          .font(.title2)
-          .fontWeight(.bold)
-          .foregroundStyle(.bootstrapPurple)
-          .padding(.top, 80)
-          .padding(.bottom, 16)
-
-        let description = String(forKey: "\(sectionType.rawValue)_text", language: language)
-        Text(markdown: description)
-          .horizontalAlignment(description.count > 100 ? .leading : .center)
-          .font(.body)
-          .foregroundStyle(.dimGray)
-      }
-    }
+    let sectionTypes = CodeOfConductSectionType.allCases
+    SectionListComponent(title: title, dataSource: sectionTypes, language: language)
 
     MainFooterWithBackground(language: language)
       .margin(.top, 160)
