+++ Website/Sources/Pages/FAQ.swift
@@ -1,6 +1,6 @@
 import Ignite
 
-enum FAQSectionType: String, CaseIterable {
+enum FAQSectionType: String, StringEnum {
   case aboutTrySwift = "about_try_swift"
   case conferenceLanguages = "conference_languages"
   case ticketCancellationPolicy = "ticket_cancellation_policy"
@@ -27,37 +27,13 @@ struct FAQ: StaticLayout {
   var body: some HTML {
     MainNavigationBar(path: generatePath(language:), language: language)
 
-    Text(String(forKey: "faq", language: language))
-      .horizontalAlignment(.center)
-      .font(.title1)
-      .fontWeight(.bold)
-      .foregroundStyle(.bootstrapPurple)
-      .padding(.top, 140)
-
     let sectionTypes = FAQSectionType.allCases.filter { section in
       switch language {
       case .ja: section != .visaSupport
       case .en: true
       }
     }
-
-    ForEach(sectionTypes) { sectionType in
-      Section {
-        Text(String(forKey: sectionType.rawValue, language: language))
-          .horizontalAlignment(.center)
-          .font(.title2)
-          .fontWeight(.bold)
-          .foregroundStyle(.bootstrapPurple)
-          .padding(.top, 80)
-          .padding(.bottom, 16)
-
-        let description = String(forKey: "\(sectionType.rawValue)_text", language: language)
-        Text(markdown: description)
-          .horizontalAlignment(description.count > 100 ? .leading : .center)
-          .font(.body)
-          .foregroundStyle(.dimGray)
-      }
-    }
+    SectionListComponent(title: title, dataSource: sectionTypes, language: language)
 
     MainFooterWithBackground(language: language)
       .margin(.top, 160)
